Adjust test to cover whole crate idl parsing
